Set TLSConfig.InsecureSkipVerify from GenTLSConfig
Since tls parser is common, it will be possible to set the "insecure"
config in all places where we use TLS (client connections, gateways).
It is also possible for people embedding NATS to simply pass their
own TLSConfig (with the flag set).
So instead, the server will reject that flag for all but Cluster
config.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>